+++ priv/storybook/components/input.story.exs
@@ -257,12 +257,111 @@ defmodule Storybook.Components.Input do
             },
             slots: [
               """
-              <:description>Tell us about yourself.</:description>
+              <:addon_left>Tell us about yourself.</:addon_left>
+              """
+            ]
+          }
+        ]
+      },
+      %VariationGroup{
+        id: :addons,
+        variations: [
+          %Variation{
+            id: :addon_left,
+            attributes: %{
+              type: "text",
+              label: "Left",
+              placeholder: "Some text"
+            },
+            slots: [
+              """
+              <:addon_left>
+                <Doggo.icon>
+                  #{mail_icon()}
+                </Doggo.icon>
+              </:addon_left>
+              """
+            ]
+          },
+          %Variation{
+            id: :addon_right,
+            attributes: %{
+              type: "text",
+              label: "Right",
+              placeholder: "Some text"
+            },
+            slots: [
+              """
+              <:addon_right>
+                <Doggo.icon>
+                  #{mail_icon()}
+                </Doggo.icon>
+              </:addon_right>
+              """
+            ]
+          },
+          %Variation{
+            id: :addon_left_and_right,
+            attributes: %{
+              type: "text",
+              label: "Left and right",
+              placeholder: "Some text"
+            },
+            slots: [
+              """
+              <:addon_left>
+                <Doggo.icon>
+                  #{mail_icon()}
+                </Doggo.icon>
+              </:addon_left>
+              <:addon_right>
+                <Doggo.icon>
+                  #{check_icon()}
+                </Doggo.icon>
+              </:addon_right>
               """
             ]
           }
         ]
       }
     ]
+  end
+
+  defp mail_icon do
+    """
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      stroke-width="2"
+      stroke-linecap="round"
+      stroke-linejoin="round"
+      class="lucide lucide-mail"
+    >
+      <rect width="20" height="16" x="2" y="4" rx="2" /><path d="m22 7-8.97 5.7a1.94 1.94 0 0 1-2.06 0L2 7" />
+    </svg>
+    """
+  end
+
+  defp check_icon do
+    """
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      stroke-width="2"
+      stroke-linecap="round"
+      stroke-linejoin="round"
+      class="lucide lucide-check"
+    >
+      <path d="M20 6 9 17l-5-5" />
+    </svg>
+    """
   end
 end