+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - New component: `Doggo.tag/1`.
 - Allow to visually hide labels.
 - Support autocomplete via `<datalist>` in `input/1` component.
+- Add `addon_left` and `addon_right` slots to `input/1` component.
 - Added `gettext` attribute to `Doggo.input/1`, giving you the choice to
   set the Gettext module locally.